Pretty jank "ancestor" detector
With a small amount of local branches, this does a decent job at
figuring out which branch a given commit/branch is based off. This isn't
an exact science because git branches are transient, but it seems
helpful/useful enough to Do What I Mean most of the time, even if I
don't even know what that is.

Based on: https://gist.github.com/joechrysler/6073741